Clean up of TC-3245 per andre's feedback
